Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==2.14.0
groq==0.13.0
agno==0.0.55
Pillow==11.0.0
google-generativeai==0.8.3
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
pydantic==2.10.4
